Extended functionality
Main feature is the possibility before the overwriting of the telephone book on the FRitz!Box that changes there could be saved as VCF file to a deisgnated email adress.
For this it is necessary to add PHP mailer and vCard to the vendors. In addition, PHP SOAP must be installed.

Beside of that a few minor improvements/changes e.g. took place in the run command for a extended user interaction